Implement getAttackers for a board square
no tests yet, so it might not actually work. This is necessary to make sure castling (or really any move) is possible.

Simplify getAttackers

look for queens while scanning rook and bishop moves, instead of as a separate step. Save's a queen's worth of move generation, which was redundant with rook and bishop.